+++ test/runtests.jl
@@ -57,9 +57,12 @@ include("data.jl")
 include("mrsp.jl")
 include("rop.jl")
 include("restoration_redispatch.jl")
-include("iter.jl")
 
-include("util.jl")
+include("final.jl")
+include("utilization.jl")
+include("rrr.jl")
+include("rad.jl")
+
 
 end